breaking up handler.js
